llm-content-eval/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.3.2
python-dotenv==1.0.0
pydantic==2.5.3
jinja2==3.1.3
python-multipart==0.0.6
aiofiles==23.2.1